Update patch set 1
Patch Set 1:

yes, sorry, i didnt mention. close-up is present and your code is done well so it works. Only the other mode is missing. We can get back to this later, so don't abandon it yet. Once things are settled generally, i want to try the new blobs too with a thorough test.

Patch-set: 1