Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I would not add that. It gives the user the false impression that the things they can adjust should work. But we are limited to the green and amber color by hardware and the blink rate is hardcoded in the kernel driver.

Please correct me if I'm wrong.

Patch-set: 1