Update patch set 1
Patch Set 1:

yes, you're right: partially. 

configurable colors are only green and amber, the other colors are not for our device and therefore a little bit confusing.
hardcoded blink rate: correct. btw: where is this given...? when in fw/base/...../config.xml: is it possible to leave this entry empty, so that it may use the values given in config menu?

at least it's possible now to configure or change green / red, per app. one can add apps, too. and battery led is "off switchable" (enerving for me, personally)

Patch-set: 1