Update patch set 3
Patch Set 3:

Huh? I messed up somewhere. I didn't want to update this change. Maybe it happened automatically because it was a dependency of another change.

I came here to say that with the other liblights change it is now possible to use a long blink mode for the amber LED. It will be used if either the pulse length or blink frequency is set to long or higher.

Implementing the other blink modes in the kernel is possible but I don't see any advantage in this at the moment.

So I'm still somewhat against this change. I know that I played around with exactly this in previous CM versions and wondered why it had no effect.

The battery led change is okay with me.

Also what do you mean with: "btw: where is this given...? when in fw/base/...../config.xml: is it possible to leave this entry empty, so that it may use the values given in config menu?"

Patch-set: 3