Update patch set 3
Patch Set 3:

in config.xml of frameworks are some addtitional bools:

config_defaultNotificationLedOff
config_defaultNotificationLedOn
config_notificationsBatteryLedOff

maybe 
config_multiColorBatteryLed
affects the colors-list in a positive way, or better: the way we need it with, only the colors our device can offer. have to try this, but can earliest do this on coming thursday.

Patch-set: 3