Update patch set 1
Patch Set 1: No score

jenkins uses different sources, especially for recovery: this was the case for cm9 and cm10 too. private recovery builders won't be happy, when their devices switch off due to low battery and they can't charge their device.

but on the other side: who is building recoveries...? the ones who kno the necessary stuff can do this themselves, so:

this isn't really "needed", yes :) but again, as for some of my other patches: a "nice to have" :)

Patch-set: 1