Update patch set 5
Patch Set 5:

rebased, still waiting for this:
"Here's the review for the option to Hide Speed/Length of blinking. I think the color selection is okay with the palette view even if it's not perfectly showing the correct color on the led, user can set a red or green color with the color picker still, so there's meaning for the color picker.

http://review.cyanogenmod.org/#/c/33537/3"

Patch-set: 5