Update patch set 2
Patch Set 2:

actually, without a correct ratio for preview this is a bit wonky to have. As seadersn took these screenshots, you can see the problem with this patch at the moment. You won't actually see the correct preview of the picture you will take in widescreen mode:
http://www.abload.de/image.php?img=saux8.jpg http://www.abload.de/image.php?img=47uid.jpg
there's some patches for this to give a true preview:
https://github.com/maxwen/android_packages_apps_Camera/commit/e4f33d5f771e5e693a113cdc9fabfac0fd1966ff
https://github.com/maxwen/android_packages_apps_Camera/commit/5aa8681355edc067f065904712c65d4253cffbca
https://github.com/maxwen/android_packages_apps_Camera/commit/65e70013413bf4a95d6856ef0e2f30052d39ea17

it would be better to have them incorporated later, along with this submitted patch.
Anyway, im fine with this overlaying, i dont think its a big issue that the preview is not "true" :)

Patch-set: 2