Update patch set 2
Patch Set 2:

what i dont get, why is there an option for true preview? it doesnt work for 4:3 well or for what reason the user must choose for true preview? It's not a question of option to show the preview of the image that will actually be taken, no?

Patch-set: 2