Update patch set 2
Patch Set 2:

if taking it serious: no, in my opinion. true preview should be standard imo. but some ppls are happy with the non-true preview, they prefer this over the true preview, even when parts of the resulting pic aren't shown. so, for these ppls, it's the best solution to have this as an option.

my thoughts about it :D

Patch-set: 2