Update patch set 2
Patch Set 2:

without the preview patches, at the moment how does the preview look like for 16:9?
For 4:3 i see that the image is stretched and the photo aspect ration will be 4:3, while you see the preview in 16:9.

I haven't tried the new resolutions yet, but what happens if you set it to 6mpixel 16:9? You get the 4:3 full sensor image in 16:9 but the image stored will be cropped to 16:9 smaller size, right, so effectively the image you see on the preview will lose the cropped part?
It's interesting, because i guess that with the true preview patch, for 4:3 you will get a 4:3 preview that only partially fits the screen, right?
Now after thinking more about this, i start to think that while the true preview is a nice thing, it would kindof work inconvenient on 4:3 ratios....
i'll think more about this, because this seems to be a more delicate matter than first it seemed to me :)

Patch-set: 2