Update patch set 1
Patch Set 1:

Yes, endeavoru is the HTC One X and enrc2b is the X+. M7 is the HTC One. 

It's nice to see that this is working and the better place would be tegra3-common. In case it causes trouble on enrc2b we can still add it to endeavoru later.

Patch-set: 1