Update patch set 2
Patch Set 2:

After going through 10 reboots with the 2013-12-24 from XDA and my build with 2014-01-08 + this commit, it looks like the behavior is about the same on both.

Roughly 30% of the time Bluetooth will fail to turn on, and enabling it results in the errors I noted earlier.  The other 70% of the time it works.

Since there is no change from the XDA snapshot posted, I recommend commiting, so at least the system has a chance of working as-is. :-)

Patch-set: 2