endeavoru: Don't start hdmid
Currently the hdmi daemon starts at boot.  It appears this daemon
both doesn't work, and has the potential to consume vast quantities
of RAM (it gets up to 90MB resident on my HOX after a few days
of uptime...although that's apparently not comment).

Since it's non-functional, don't bother starting it. :-)

Change-Id: If5a58e77afcb4cb2e8edb9b194bfedf2ee4757cb